# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "技術トレンド キャッチアップ API",
    "description": "Qiita/Zenn/Hacker News/ArXivから自動収集した技術記事フィード、Gemini選定のAIツールピックアップ、音声ポッドキャストを取得できる読み取り専用の公開API。認証不要・無料。Claude Code/Claude Desktopからは同じデータをMCPサーバー（/api/mcp）経由でも利用できます。",
    "version": "1.0.0"
  },
  "servers": [{ "url": "https://dev-current.vercel.app" }],
  "paths": {
    "/api/feed": {
      "get": {
        "operationId": "getFeed",
        "summary": "技術記事フィードを取得",
        "description": "日別または週間（直近7日間、JST基準）の注目ピックアップと通常記事一覧を返します。",
        "parameters": [
          { "name": "period", "in": "query", "schema": { "type": "string", "enum": ["day", "week"], "default": "week" }, "description": "day=指定日のみ, week=直近7日間" },
          { "name": "date", "in": "query", "schema": { "type": "string", "format": "date" }, "description": "YYYY-MM-DD（JST）。未指定なら今日" },
          { "name": "country", "in": "query", "schema": { "type": "string" }, "description": "発信国で絞り込む（例: アメリカ）" }
        ],
        "responses": {
          "200": {
            "description": "フィード取得成功",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/FeedResponse" } } }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "FeedArticle": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "url": { "type": "string" },
          "title": { "type": ["string", "null"] },
          "sourceType": { "type": "string", "enum": ["QIITA", "ZENN", "HACKER_NEWS", "ARXIV", "AI_TOOL_PICK", "USER_SUBMITTED"] },
          "publishedAt": { "type": ["string", "null"], "format": "date-time" },
          "country": { "type": ["string", "null"] },
          "engagementScore": { "type": ["integer", "null"] },
          "hasInsight": { "type": "boolean" }
        }
      },
      "FeedResponse": {
        "type": "object",
        "properties": {
          "period": { "type": "string", "enum": ["day", "week"] },
          "date": { "type": "string" },
          "featured": { "type": "array", "items": { "$ref": "#/components/schemas/FeedArticle" } },
          "regular": { "type": "array", "items": { "$ref": "#/components/schemas/FeedArticle" } },
          "regularOffset": { "type": "integer" },
          "regularCount": { "type": "integer" },
          "regularTotal": { "type": "integer" },
          "hasMore": { "type": "boolean" }
        }
      },
...
    }
  }
}